Docker build & clean up audio code, WIP better spectrogram memory storage

+++ src/app/memusage.h
@@ -1,13 +1,14 @@
 #ifndef REFORMANT_MEMUSAGE_H
 #define REFORMANT_MEMUSAGE_H
 
+#include <cstddef>
 #include <cstdint>
 
 namespace reformant {
 uint64_t memoryUsage();
 }  // namespace reformant
 
-inline uint64_t operator"" _u64(uint64_t x) { return x; }
-inline size_t operator"" _sz(size_t x) { return x; }
+inline uint64_t operator"" _u64(unsigned long long int x) { return x; }
+inline size_t operator"" _sz(unsigned long long int x) { return x; }
 
 #endif  // REFORMANT_MEMUSAGE_H
